Move observers into their own files
