+++ src/app/wishlists/reducer.ts
@@ -4,8 +4,6 @@ import { ActionType, getType } from 'typesafe-actions';
 import { getInventoryWishListRolls } from './wishlists';
 import { RootState } from 'app/store/types';
 import _ from 'lodash';
-import { observeStore } from '../utils/redux-utils';
-import { set } from 'idb-keyval';
 import { WishListAndInfo } from './types';
 import { createSelector } from 'reselect';
 import { storesSelector } from '../inventory/selectors';
@@ -78,17 +76,3 @@ export const wishLists: Reducer<WishListsState, WishListAction> = (
       return state;
   }
 };
-
-export function saveWishListToIndexedDB() {
-  return observeStore(
-    (state) => state.wishLists,
-    (_, nextState) => {
-      if (nextState.loaded) {
-        set('wishlist', {
-          wishListAndInfo: nextState.wishListAndInfo,
-          lastFetched: nextState.lastFetched,
-        });
-      }
-    }
-  );
-}
